requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.4
Markdown==3.3.6
mysqlclient==2.2.0
Pillow==10.1.0
pytz==2021.3
requests==2.31.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==2.0.7
zipp==3.6.0


# to use this requirement.txt run this below cmmd

# pip install -r requirements.txt
